修改了几个实体类的结构，包括增加字段跟修改数据类型，xml文件也修改了对应关系

+++ src/main/java/com/example/mybatisplus/model/domain/Record.java
@@ -33,7 +33,7 @@ public class Record extends Model<Record> {
 
     private Integer comId;
 
-    private Integer number;
+    private Double number;
 
     private Boolean isDeleted;
